requirements.txt
aiodns==3.0.0
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
aleph-alpha-client==2.16.0
altair==4.2.2
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
blinker==1.5
boto3==1.26.81
botocore==1.29.81
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
dataclasses-json==0.5.7
decorator==5.1.1
deeplake==3.2.12
dill==0.3.6
dnspython==2.3.0
entrypoints==0.4
Faker==17.4.0
favicon==0.7.0
fonttools==4.38.0
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
htbuilder==0.6.1
hub==3.0.1
humbug==0.2.8
idna==3.4
importlib-metadata==6.0.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
kiwisolver==1.4.4
langchain==0.0.96
loguru==0.6.0
lxml==4.9.2
Markdown==3.4.1
markdown-it-py==2.2.0
markdownlit==0.0.5
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.0
mdurl==0.1.2
more-itertools==9.1.0
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
numcodecs==0.11.0
numpy==1.24.2
openai==0.26.5
packaging==23.0
pandas==1.5.3
pathos==0.3.0
Pillow==9.4.0
pinecone-client==2.2.1
pox==0.3.2
ppft==1.7.6.6
protobuf==3.20.3
pyarrow==11.0.0
pycares==4.3.0
pycparser==2.21
pydantic==1.10.5
pydeck==0.8.0
Pygments==2.14.0
PyJWT==2.6.0
pymdown-extensions==9.9.2
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.2
rich==13.3.1
s3transfer==0.6.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
soupsieve==2.4
SQLAlchemy==1.4.46
st-annotated-text==3.0.0
streamlit==1.19.0
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.4
streamlit-chat==0.0.2.1
streamlit-embedcode==0.1.2
streamlit-extras==0.2.6
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.3
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
tenacity==8.2.2
tokenizers==0.13.2
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.20.0
yarl==1.8.2
zipp==3.15.0
